# change labelset comparison in promql engine to avoid false positive during detection of duplicates