--- a/promql/functions.go
+++ b/promql/functions.go
@@ -14,7 +14,6 @@
 package promql
 
 import (
-	"bytes"
 	"math"
 	"regexp"
 	"sort"
@@ -599,8 +598,7 @@ func funcPredictLinear(vals []parser.Value, args parser.Expressions, enh *EvalNo
 func funcHistogramQuantile(vals []parser.Value, args parser.Expressions, enh *EvalNodeHelper) Vector {
 	q := vals[0].(Vector)[0].V
 	inVec := vals[1].(Vector)
-	b := bytes.Buffer{}
-	sigf := signatureFunc(false, &b, excludedLabels...)
+	sigf := signatureFunc(false, &enh.lblBuf, excludedLabels...)
 
 	if enh.signatureToMetricWithBuckets == nil {
 		enh.signatureToMetricWithBuckets = map[string]*metricWithBuckets{}