… pkg/labels/labels_test.go
… promql/testdata/collision.test