This PR changes from using hashes to compare labelsets to see if labelsets on each side of a promql query are dupicated to just using strings of `labels.Labels`.

Signed-off-by: Callum Styan <callumstyan@gmail.com>