@roidelapluie 2020-03-26
Should we also update Hash() to use this?