@brian-brazil 2020-03-26
We should look at eliminating the hashing, rather than changing the function.