@cstyan 2020-03-26
> We should look at eliminating the hashing, rather than changing the function.

Can you elaborate as to how that would work? Just compare labels.Labels and ensure there's at least one unique label on one side?

Also, are you opposed to changing the hashing function where I have in order to fix the issue in the short term, assuming we move to go1.14?