@brian-brazil 2020-03-26
Changing the hash function only changes what the collisions are.