@cstyan 2020-03-26
Given that this is what Go uses internally for maps I don't see the issue with merging this, given that it should be less prone to collisions that what we have now, and implementing something better that doesn't use hashes in the future.