@brian-brazil 2020-03-26
But Go has collision handling, so making presumptions about its collision properties merely because Go uses it seems a bit of a leap.