@roidelapluie 2020-03-26
> But Go has collision handling, so making presumptions about its collision properties merely because Go uses it seems a bit of a leap.

 The hash functions are collision-resistant but not cryptographically secure. (See crypto/sha256 and crypto/sha512 for cryptographic use.) 
https://golang.org/pkg/hash/maphash/