@cstyan 2020-03-26
It's been 9 months since the original issue was filed and no one's come forward with an implementation other than replacing the hash function. Should we not just move forward with one of those and file an issue for the other implementation?