@roidelapluie 2020-03-26
I'd be in favor of using maphash as it is documented as collision resistant (which is the issue here). I'd still like to see it applied everywhere then.