@brian-brazil 2020-03-27
Just because noone has yet created the correct fix doesn't mean that we should implement an incorrect fix that will itself cause unknown problems. Collisions are inevitable.