@pstibrany 2020-03-27
> Just because noone has yet created the correct fix doesn't mean that we should implement an incorrect fix that will itself cause unknown problems. Collisions are inevitable.

Using a collision resistant hash function means that you would need to create *large* number of hash values to get a guaranteed collision (think 7x10^9 for 50% collision chance for 64-bit hash space). By that time, you will run out of memory first. Probability of getting a collision with random values is much lower than with "traditional" hash function. That still sounds like an improvement to me.