@brian-brazil 2020-03-27
The rule of thumb is the square root for a 50% collision chance. Given the number of users we have, we're going to have collisions.