@pstibrany 2020-03-27
> The rule of thumb is the square root for a 50% collision chance. Given the number of users we have, we're going to have collisions.

Yeah, but square root of 64-bit hash space is still 5.1x10^9 values before you get there. (I misremembered, and wrote 7 above).