@beorn7 2020-03-27
With no PR in sight to implement collision detection, I'm all for a simple change that makes it less likely to run into a hash collision.

However, is the new `maphash` in Go 1.14 really more collision-resistant than the `xxhash` we are using now? [xxhash](https://cyan4973.github.io/xxHash/) claims to score perfectly in the [SMHasher](https://github.com/aappleby/smhasher/wiki/SMHasher) test. I don't know if that means it is close to the above mentioned square-root probability. And of course, we cannot really say how close the `maphash` hash function is (I suspect the actual function used might depend on the hardware).

Has anybody any evidence how the two compare?