@cstyan 2020-03-27
@brian-brazil so it sounds like the "correct solution" is to replace the maps in the engine that are `uint64 -> labels.Labels` with maps of `labels.String() -> labels.Labels` or something similar. Without hashing the label set ourselves we could just rely on the runtimes collision handling for maps.