@brian-brazil 2020-03-27
Yeah, something like that would be the way to do it. The main question is what the CPU hit will be.