@cstyan 2020-03-27
Okay I think I have something working via maps, I'll clean it up and push in a bit.

@brian-brazil AFACIT for `TestEvaluations` the test 
`eval instant at 1m absent_over_time(rate(nonexistant[5m])[5m:])` from `testdata/functions.test` fails even on master. Is that something we should be looking into?