@roidelapluie 2020-03-27
```
~/go/src/github.com/prometheus/prometheus/promql $ git rev-parse HEAD
0c4ec8d9dd1c64363dce59bed45a759f7f2149df
~/go/src/github.com/prometheus/prometheus/promql $ go test
PASS
ok      github.com/prometheus/prometheus/promql 0.361s
```

it passes for me