@cstyan 2020-03-27
It only seems to fail if you specifically run `TestEvaluations`

```
prometheus/promql branch  master [$?] go v1.14.1 ➜ git rev-parse HEAD
0c4ec8d9dd1c64363dce59bed45a759f7f2149df

prometheus/promql branch  master [$?] go v1.14.1 ➜ go test -run TestEvaluations -v
=== RUN   TestEvaluations
    TestEvaluations: promql_test.go:31: unexpected error: error evaluating query "absent_over_time(rate(nonexistant[5m])[5m:])" (line 590): unexpected error: runtime error: integer divide by zero
--- FAIL: TestEvaluations (0.14s)
FAIL
exit status 1
FAIL	github.com/prometheus/prometheus/promql	0.143s

prometheus/promql branch  master [$?] go v1.14.1 ➜ go test ./...
ok  	github.com/prometheus/prometheus/promql	(cached)
ok  	github.com/prometheus/prometheus/promql/parser	(cached)```