@cstyan 2020-04-13
https://gist.github.com/cstyan/d80b13d2c5e62447abbdcf3c6bb6417c @brian-brazil the alloc/bytes benchmarks don't look great. I'll let you comment on whether the changes are acceptable based on the benchmarks, or if there's something I'm doing in an inefficient way. I've made the changes to signatureFunc and various maps based on my understanding of our discussion in earlier comments.