@brian-brazil 2020-04-14
I'm more concerned that it's causing those queries to take twice as long. Are there any easy wins for optimsation?