@cstyan 2020-04-23
Made some optimizations in the latest commit, mostly around using a `sync.Pool` in for the key in `resultMetric` and for the buffers in `labels.go` `String`. 

Here are the benchmark comparisons of master vs the latest commit and master vs the previous commit: https://gist.github.com/cstyan/20b58011ca95694fd56ec89d49000729

**Highlights:**

Worst ns/op in previous commit 
```
benchmark                                                                                                             old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                          868686865      2352684729     +170.83%
```

And worst ns/op in latest commit
```
benchmark                                                                                                             old ns/op      new ns/op      delta
BenchmarkRangeQuery/expr=a_ten_-_b_ten,steps=10000-8                                                                  80429636       130171151      +61.84%
```

Worst allocs in previous commit
```
benchmark                                                                                                             old allocs     new allocs     delta
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                          140992         21150098       +14900.92%
```

Worst allocs in latest commit
```
benchmark                                                                                                             old allocs     new allocs     delta
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                          140992         5176815        +3571.71%
```

And finally worst bytes/op in previous commit
```
benchmark                                                                                                             old bytes     new bytes     delta
BenchmarkRangeQuery/expr=a_hundred_-_b_hundred,steps=10000-8                                                          22621664      598825712     +2547.13%
```

And worst bytes/op in latest commit
```
benchmark                                                                                                             old bytes     new bytes     delta
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1000-8                                    7728456       86935588      +1024.88%
```
This is a different expression though, that expression in the previous commit
```
BenchmarkRangeQuery/expr=histogram_quantile(0.9,_rate(h_hundred[5m])),steps=1000-8                                    7728456       138916932     +1697.47%
```

Thanks to @cyriltovena and @peterbourgon for their help with benchmarks and potential optimizations